sql: split out pure module
This keeps all the code for statement purification in one place, and
makes for a nice place to document the approach.
